+++ clientele/settings.py
@@ -1,12 +1,6 @@
 import platform
-from os.path import abspath, dirname
 
-from jinja2 import Environment, PackageLoader
-
-TEMPLATES_ROOT = dirname(dirname(abspath(__file__))) + "/clientele/templates/"
 VERSION = "0.7.0"
-
-templates = Environment(loader=PackageLoader("clientele", "templates"))
 
 
 def split_ver():